Refactor dashboard to separate component
To get ready for switching pages
